# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Couture Dynamic Pro — API métier (Phase 4)
  version: "1.0"
  description: >
    Contrat des endpoints REST versionnés (collections, catalogues, pages,
    produits, variantes, panier, commandes). Les endpoints /api/auth et
    /api/admin/users (Phase 3) et les paiements (Phase 7) ne sont pas
    couverts par ce document.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cdp_session
...
paths:
...
  /produits/{slug}:
    get:
      summary: Détail d'un produit avec ses variantes (ACTIVE uniquement pour le public, RG5)
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Non trouvé / non actif }
  /produits/{id}:
    patch:
      summary: Mettre à jour un produit ; passage à ACTIVE exige une variante ACTIVE (RG5)
      description: Rôle autorisé — ADMIN uniquement.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "400": { description: "RG5 non satisfait (NO_ACTIVE_VARIANT)" }
  /produits/{id}/variantes:
    post:
      summary: Ajouter une variante (prix/stock >= 0, RG5, contrainte DB)
      description: Rôle autorisé — ADMIN uniquement.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "201": { description: Créée }
        "409": { description: SKU ou combinaison taille/couleur déjà utilisés }
  /produits/{id}/variantes/{varianteId}:
    patch:
      summary: Mettre à jour une variante (prix, stock, statut)
      description: Rôle autorisé — ADMIN uniquement.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: varianteId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...